Add refactor database
